Fix regression in rendering and turn on reflections for renderer tests.
The HD renderer did draw reflections in a different way, because
the index for the sinus table was computed differently as compared
to the original renderer.
Fixed this and the renderer tests that compare HD and original have
reflections now enabled.

(cherry picked from commit 5c5db653449f86214dce0b9899224348a764588d)